Rework attachments display
