+++ packages/api/src/localization/en-US.json
@@ -126,7 +126,14 @@
   "TEXT_INPUT_TELEPHON_KEYPAD_BUTTON_ALT": "Dialpad",
   "TEXT_INPUT_DROP_ZONE": "Drop files",
   "TEXT_INPUT_LENGTH_EXCEEDED_ALT": "Message lengths exceeded",
-  "TEXT_INPUT_ATTACHMENTS": "attachments",
+  "TEXT_INPUT_ATTACHMENTS_FEW": "$1 attachments",
+  "_TEXT_INPUT_ATTACHMENTS_FEW.comment": "$1 is the number of attachments. This is for plural rule of \"few\".",
+  "TEXT_INPUT_ATTACHMENTS_MANY": "$1 attachments",
+  "_TEXT_INPUT_ATTACHMENTS_MANY.comment": "$1 is the number of attachments. This is for plural rule of \"many\".",
+  "TEXT_INPUT_ATTACHMENTS_OTHER": "$1 attachments",
+  "_TEXT_INPUT_ATTACHMENTS_OTHER.comment": "$1 is the number of attachments. This is for plural rule of \"other\".",
+  "TEXT_INPUT_ATTACHMENTS_TWO": "$1 attachments",
+  "_TEXT_INPUT_ATTACHMENTS_TWO.comment": "$1 is the number of attachments. This is for plural rule of \"two\".",
   "TOAST_ACCORDION_FEW": "$1 Notifications: Click here to see details",
   "_TOAST_ACCORDION_FEW.comment": "$1 is the number of notifications. This is for plural rule of \"few\".",
   "TOAST_ACCORDION_MANY": "$1 Notifications: Click here to see details",
